requirements.txt
Flask==2.0.0
certifi==2023.7.22
chardet==5.2.0
click==8.0.0
Flask-JWT-Extended==4.4.4
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
flask-swagger-ui==4.11.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.29.0
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing-extensions==4.8.0
urllib3==2.2.3
Werkzeug==2.0.0
WTForms==3.0.1
kafka-python
six>=1.15.0
